move runAgent into agent file
